Refactor: Projectile to module
