+++ CondaRTS.py
@@ -25,6 +25,7 @@ from src.game_object import GameObject
 from src.geometry import is_valid_building_position, snap_to_grid
 from src.iron_field import IronField
 from src.particle import Particle
+from src.projectile import Projectile
 from src.shapes import draw_progress_bar
 
 if TYPE_CHECKING:
@@ -777,73 +778,6 @@ class Turret(Building):
         self.image.set_alpha(
             int(255 * self.construction_progress / self.CONSTRUCTION_TIME)
         )
-
-
-class Projectile(pg.sprite.Sprite):
-    SPEED: float = 6
-
-    def __init__(
-        self, x: float, y: float, target_unit: GameObject, damage: int, team: Team
-    ) -> None:
-        super().__init__()
-        self.image: pg.Surface = pg.Surface((10, 5), pg.SRCALPHA)
-        self.rect: pg.Rect = self.image.get_rect(center=(x, y))
-        self.target_unit = target_unit
-        self.damage = damage
-        self.team = team
-        self.particle_timer = 2
-
-        pg.draw.ellipse(self.image, (255, 200, 0), (0, 0, 10, 5))
-
-    def update(self) -> None:
-        if self.target_unit and self.target_unit.health > 0:
-            dx, dy = (
-                self.target_unit.rect.centerx - self.rect.centerx,
-                self.target_unit.rect.centery - self.rect.centery,
-            )
-            dist = math.sqrt(dx**2 + dy**2)
-            if dist > 3:
-                angle = math.atan2(dy, dx)
-                self.image = pg.transform.rotate(
-                    pg.Surface((10, 5), pg.SRCALPHA), -math.degrees(angle)
-                )
-                pg.draw.ellipse(self.image, (255, 200, 0), (0, 0, 10, 5))
-                self.rect.x += self.SPEED * math.cos(angle)
-                self.rect.y += self.SPEED * math.sin(angle)
-                if self.particle_timer <= 0:
-                    particles.add(
-                        Particle(
-                            self.rect.centerx,
-                            self.rect.centery,
-                            -math.cos(angle) * random.uniform(0.5, 1.5),
-                            -math.sin(angle) * random.uniform(0.5, 1.5),
-                            5,
-                            (255, 255, 150),
-                            15,
-                        )
-                    )
-                    self.particle_timer = 2
-                else:
-                    self.particle_timer -= 1
-            else:
-                self.kill()
-                for _ in range(5):
-                    particles.add(
-                        Particle(
-                            self.rect.centerx,
-                            self.rect.centery,
-                            random.uniform(-2, 2),
-                            random.uniform(-2, 2),
-                            6,
-                            (255, 100, 0),
-                            15,
-                        )
-                    )  # Orange explosion
-        else:
-            self.kill()
-
-    def draw(self, screen: pg.Surface, camera: Camera) -> None:
-        screen.blit(self.image, camera.apply(self.rect).topleft)
 
 
 @dataclass(kw_only=True)
